# api/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Public API
  version: v1
  description: External gateway for sessions, SSE events, cancel, and HITL. Token auth (not Hands HMAC).
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      security: []
      responses:
        "200":
          description: ok
  /v1/sessions:
    post:
      summary: Create session and start a run
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "202":
          description: accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "401":
          description: unauthorized
  /v1/sessions/{id}:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
  /v1/sessions/{id}/events:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: SSE stream of protocol.Event
  /v1/sessions/{id}/cancel:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "202":
          description: cancel requested
  /v1/sessions/{id}/hitl:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HITLRequest"
      responses:
        "202":
          description: decision queued
  /v1/sessions/{id}/messages:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "202":
          description: accepted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    CreateSessionRequest:
      type: object
      required: [message]
      properties:
        message: {type: string}
        workspace: {type: string}
    CreateSessionResponse:
      type: object
      properties:
        session_id: {type: string}
    SessionStatus:
      type: object
      properties:
        session_id: {type: string}
        running: {type: boolean}
        final_text: {type: string}
        error: {type: string}
    HITLRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, deny]
    Event:
      type: object
      properties:
        type: {type: string}
        session_id: {type: string}
        turn_id: {type: string}
        data: {type: object}
        at: {type: string, format: date-time}